Import go images.
This imports the go base image and 'go-wrapper' from
gcb-catalog/go-images at a2f95002fc. Note that this adopts
a multi-version go image, with a custom entrypoint that can be
used to select the go version to use; instead of having multiple
images per go version. This was proposed in
go/multi-version-cicd-component-images.

Change-Id: Idfa5055462eaf07d25781ba32e1ffe0baeef0990