Add separate setup_*.sh script for setting up a toolchain.
Here, we split the entrypoint.sh into a separate script that can
be used directly (by sourcing) to set the PATH and other
toolchain specific environment variables. The images still work
as before, with a default entrypoint.sh, but can now be used
in contexts where the entrytpoint is over-written (e.g script
mode in Tekton).

BUG=318718506

Change-Id: I68f0631824ec93912517bbb1069be060f6d0a7f1