Add a presubmit script.
This will simply look at the current commit and iterate over each
image folder to run `make build test`. This might be slow for
the case when files in all the folders are changed, so we can
improve this over time, but for now this will significantly help
vs. running all build/test for each change.

Change-Id: I1afacbeba64733c31274b3129855006f314ad616