Use (google) go builder images.
I think we should just be using the Google provided image directly. It introduces one less dependency
(albiet local) into the build graph and I think makes it a bit more clear how to build a go binary, then use it in another image.

This came up in our meeting with the Longshore folks today too, so I think the pattern is a bit more clear now.

We will need to periodically bump the go version used, but that should just be a find/replace.

Change-Id: I21bb74ec121ac72d0ec601755b465f79767ebb11