chore: Some updates to our pre-commit config.
Here we specify the default stages to install hooks for.
This means folks can still just run `pre-commit install` once,
without specifying all the stages explicitly.

We also setup the `commitzen-branch` stage, and set it to be
run in manual mode only. We will use this in the presubmit validation
that changes look good.

Change-Id: I8d5f6ad677aba2e2786caa7b27e99d4b1c641719