chore: Refactor go code into a top-level module.
Here, we start the process of refactoring the go code to a top-level
module. This will allow us to have fewer individual go.mod files
to manage (minimizing dependency updates), and allow us to start
sharing common libraries/functionality across binaries.

For now, we move all images to a sub-directory, and keep the
top-level directory looking like a regular go module. We may
want to change this in the future, but for now it allows folks
to primarily focus on writing go binaries and promotes re-use
across binaries.

Change-Id: I9234edd9b4920e7348f0f9ab6fa5c0d21df6ced7