fix: do not return error on skaffold-version
This commit fixes the unexpected error when specifying unknown "--skaffold-verion" issue identified in the bugbash.

Prior to this commit, the validateSupportedSkaffoldVersion() function
returns error when the specified version is not found. This commit do
not create this error to support extra allowed versions like
"skaffold_preview".

This is consistent with `gcloud deploy release create` command.

This commit also makes the user-facing logging msg capitalized.

Skip-Verified: Louhi test pipeline has passed but the stamp is not granted properly.

https://louhi.dev/4923718738116608/execution-detail/6650005100822528

Change-Id: I1f2fcf6fb5ea2f2a414db009d45787f9012cdd66