fix: fix cloud deploy duplicated header issue
Prior to this commit, the createTarball() function has duplicated
tarWriter.WriteHeader() function call which causes duplciated files in
the tarfile, causing cloud deploy failures.

This commit fixes the duplicated call and did small cleanup to the
function. This commit also ignores .gitignore a nd .git files when
creating the tarball, which is the default behaviour of `gcloud deploy
release create ...`

Change-Id: I9bf617662d9a2a398c185fd565e0aea0cd883e2e