chore: Update the builder image
The builder image is relatively lightweight, as it's currently only used in the GHA-on-GCB project and will likely undergo some changes before it's made any sort of 'official' image  for GCB. As such, I'm adding some cleanup to remove unused scripts and try to centralize as much as possible in the 'builder' directory as it's a likely candidate to be moved to a separate repo.

Also, I've switched to use the slim gcloud image (debian bookworm based). Which has a few implications, like no ppa for git.

Change-Id: I96e73f6e329f433a2f1907c8b10c50190891bc25